+++ api/models.go
@@ -39,6 +39,7 @@ type Series struct {
 	Title      string    `json:"title"`
 	Sub        string    `json:"sub,omitempty"`
 	Progress   int       `json:"progress"`
+	Season     int       `json:"season"`
 	Status     string    `json:"status,omitempty"`
 	StatusType string    `json:"statusType,omitempty"`
 	CreatedAt  time.Time `json:"created_at"`
@@ -50,6 +51,7 @@ type Movie struct {
 	Emoji      string    `json:"emoji,omitempty"`
 	Title      string    `json:"title"`
 	Sub        string    `json:"sub,omitempty"`
+	Genre      string    `json:"genre,omitempty"`
 	Status     string    `json:"status,omitempty"`
 	StatusType string    `json:"statusType,omitempty"`
 	CreatedAt  time.Time `json:"created_at"`
@@ -64,5 +66,6 @@ type Activity struct {
 	Who       string    `json:"who"`
 	Date      string    `json:"date,omitempty"`
 	Time      string    `json:"time,omitempty"`
+	Status    string    `json:"status,omitempty"`
 	CreatedAt time.Time `json:"created_at"`
 }